Move breed stuff to filters screen
[#158054024]
