+++ app/src/main/java/tech/pauly/findapet/shared/datastore/FilterAnimalTypeUseCase.java
@@ -2,10 +2,10 @@ package tech.pauly.findapet.shared.datastore;
 
 import tech.pauly.findapet.data.models.AnimalType;
 
-public class BreedAnimalTypeUseCase implements UseCase {
+public class FilterAnimalTypeUseCase implements UseCase {
     private AnimalType animalType;
 
-    public BreedAnimalTypeUseCase(AnimalType animalType) {
+    public FilterAnimalTypeUseCase(AnimalType animalType) {
         this.animalType = animalType;
     }
