--- app/src/main/res/layout/activity_breed.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-
-        <variable
-            name="viewModel"
-            type="tech.pauly.findapet.discover.BreedViewModel" />
-
-    </data>
-
-    <android.support.constraint.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@android:color/white"
-        android:elevation="0dp">
-
-        <android.support.v7.widget.Toolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:background="@android:color/white"
-            android:theme="@style/LightToolbarTheme"
-            app:layout_collapseMode="pin"
-            app:layout_constraintTop_toTopOf="parent"
-            app:title="@string/breed_title"
-            app:titleTextAppearance="@style/FontStrongHeader"
-            app:titleTextColor="@color/purpleStandard" />
-
-        <Button
-            android:id="@+id/button_search"
-            android:layout_width="match_parent"
-            android:layout_height="40dp"
-            android:background="@drawable/selector_large_button"
-            android:onClick="@{viewModel::saveBreed}"
-            android:text="@string/breed_search_button"
-            app:layout_constraintBottom_toBottomOf="parent" />
-
-    </android.support.constraint.ConstraintLayout>
-</layout>
