+++ app/src/main/res/layout/activity_filter.xml
@@ -37,52 +37,6 @@
             app:titleTextColor="@color/purpleStandard" />
 
         <TextView
-            android:id="@+id/breed_header"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="@dimen/standard_margin"
-            android:layout_marginStart="@dimen/standard_margin"
-            android:layout_marginTop="@dimen/standard_margin"
-            android:text="@string/breed_header"
-            android:textAppearance="@style/FontStrongHeader"
-            android:textColor="@color/textStandard"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/toolbar" />
-
-        <Button
-            android:id="@+id/button_breed_search"
-            android:layout_width="match_parent"
-            android:layout_height="40dp"
-            android:background="@drawable/selector_rounded_button"
-            android:drawableEnd="@drawable/ic_search"
-            android:layout_marginStart="@dimen/standard_margin"
-            android:layout_marginEnd="@dimen/standard_margin"
-            android:layout_marginTop="@dimen/half_margin"
-            android:elevation="4dp"
-            android:gravity="start|center_vertical"
-            android:onClick="@{viewModel::clickBreedSearch}"
-            android:paddingEnd="@dimen/half_margin"
-            android:paddingStart="@dimen/half_margin"
-            android:singleLine="true"
-            android:text="@string/filter_breed_hint"
-            android:textAppearance="@style/FontSmall"
-            android:textColor="@color/purpleVeryLight"
-            android:textOff="@string/small"
-            android:textOn="@string/small"
-            app:layout_constraintTop_toBottomOf="@id/breed_header"/>
-
-        <View
-            android:id="@+id/divider_breed"
-            android:layout_width="match_parent"
-            android:layout_height="1dp"
-            android:layout_marginEnd="@dimen/standard_margin"
-            android:layout_marginStart="@dimen/standard_margin"
-            android:layout_marginTop="@dimen/standard_margin"
-            android:background="@color/offWhite"
-            app:layout_constraintTop_toBottomOf="@+id/button_breed_search" />
-
-        <TextView
             android:id="@+id/sex_header"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
@@ -94,7 +48,7 @@
             android:textColor="@color/textStandard"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/divider_breed" />
+            app:layout_constraintTop_toBottomOf="@id/toolbar" />
 
         <RadioGroup
             android:id="@+id/sex_radio_group"
@@ -374,6 +328,42 @@
                 android:textOff="@string/extra_large"
                 android:textOn="@string/extra_large" />
         </RadioGroup>
+
+        <!--<TextView-->
+            <!--android:id="@+id/breed_header"-->
+            <!--android:layout_width="0dp"-->
+            <!--android:layout_height="wrap_content"-->
+            <!--android:layout_marginEnd="@dimen/standard_margin"-->
+            <!--android:layout_marginStart="@dimen/standard_margin"-->
+            <!--android:layout_marginTop="@dimen/standard_margin"-->
+            <!--android:text="@string/breed_header"-->
+            <!--android:textAppearance="@style/FontStrongHeader"-->
+            <!--android:textColor="@color/textStandard"-->
+            <!--app:layout_constraintEnd_toEndOf="parent"-->
+            <!--app:layout_constraintStart_toStartOf="parent"-->
+            <!--app:layout_constraintTop_toBottomOf="@id/toolbar" />-->
+
+        <!--<Button-->
+            <!--android:id="@+id/button_breed_search"-->
+            <!--android:layout_width="match_parent"-->
+            <!--android:layout_height="40dp"-->
+            <!--android:background="@drawable/selector_rounded_button"-->
+            <!--android:drawableEnd="@drawable/ic_search"-->
+            <!--android:layout_marginStart="@dimen/standard_margin"-->
+            <!--android:layout_marginEnd="@dimen/standard_margin"-->
+            <!--android:layout_marginTop="@dimen/half_margin"-->
+            <!--android:elevation="4dp"-->
+            <!--android:gravity="start|center_vertical"-->
+            <!--android:onClick="@{viewModel::clickBreedSearch}"-->
+            <!--android:paddingEnd="@dimen/half_margin"-->
+            <!--android:paddingStart="@dimen/half_margin"-->
+            <!--android:singleLine="true"-->
+            <!--android:text="@string/filter_breed_hint"-->
+            <!--android:textAppearance="@style/FontSmall"-->
+            <!--android:textColor="@color/purpleVeryLight"-->
+            <!--android:textOff="@string/small"-->
+            <!--android:textOn="@string/small"-->
+            <!--app:layout_constraintTop_toBottomOf="@id/breed_header"/>-->
 
         <Button
             android:id="@+id/button_search"
