+++ app/src/test/java/tech/pauly/findapet/discover/FilterViewModelTest.java
@@ -9,13 +9,19 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
+import java.util.Arrays;
+
 import io.reactivex.Completable;
 import io.reactivex.Single;
+import tech.pauly.findapet.data.BreedRepository;
 import tech.pauly.findapet.data.FilterRepository;
 import tech.pauly.findapet.data.models.Age;
 import tech.pauly.findapet.data.models.AnimalSize;
+import tech.pauly.findapet.data.models.AnimalType;
+import tech.pauly.findapet.data.models.BreedListResponse;
 import tech.pauly.findapet.data.models.Filter;
 import tech.pauly.findapet.data.models.Sex;
+import tech.pauly.findapet.shared.datastore.FilterAnimalTypeUseCase;
 import tech.pauly.findapet.shared.datastore.TransientDataStore;
 import tech.pauly.findapet.shared.events.ActivityEvent;
 import tech.pauly.findapet.shared.events.ViewEventBus;
@@ -23,6 +29,7 @@ import tech.pauly.findapet.shared.events.ViewEventBus;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -30,6 +37,9 @@ public class FilterViewModelTest {
 
     @Mock
     private FilterRepository filterRepository;
+
+    @Mock
+    private BreedRepository breedRepository;
 
     @Mock
     private ViewEventBus eventBus;
@@ -40,6 +50,9 @@ public class FilterViewModelTest {
     @Mock
     private Filter filter;
 
+    @Mock
+    private BreedListResponse breedListResponse;
+
     private FilterViewModel subject;
 
     @Before
@@ -47,7 +60,8 @@ public class FilterViewModelTest {
         MockitoAnnotations.initMocks(this);
         when(filterRepository.getCurrentFilter()).thenReturn(Single.just(filter));
         when(filterRepository.insertFilter(any(Filter.class))).thenReturn(Completable.complete());
-        subject = new FilterViewModel(filterRepository, eventBus, dataStore);
+        when(breedRepository.getBreedList(any(AnimalType.class))).thenReturn(Single.just(breedListResponse));
+        subject = new FilterViewModel(filterRepository, breedRepository, eventBus, dataStore);
     }
 
     @Test
@@ -155,9 +169,47 @@ public class FilterViewModelTest {
     }
 
     @Test
-    public void onClickBreedSearch_launchesBreedActivity() {
-        subject.clickBreedSearch(null);
+    public void updateBreedList_noAnimalType_doNothing() {
+        when(dataStore.get(FilterAnimalTypeUseCase.class)).thenReturn(null);
 
-        verify(eventBus).send(ActivityEvent.build(subject).startActivity(BreedActivity.class));
+        subject.updateBreedList();
+
+        verify(breedRepository, never()).getBreedList(any(AnimalType.class));
+    }
+
+    @Test
+    public void updateBreedList_getsBreedListForAnimalType() {
+        when(breedListResponse.getBreedList()).thenReturn(null);
+        setupDataStoreWithUseCase();
+
+        subject.updateBreedList();
+
+        verify(breedRepository).getBreedList(AnimalType.CAT);
+    }
+
+    @Test
+    public void updateBreedList_returnedBreedListNull_doNothing() {
+        when(breedListResponse.getBreedList()).thenReturn(null);
+        setupDataStoreWithUseCase();
+
+        subject.updateBreedList();
+
+        assertThat(subject.breedList).hasSize(0);
+    }
+
+    @Test
+    public void updateBreedList_returnedBreedListValid_showBreedList() {
+        when(breedListResponse.getBreedList()).thenReturn(Arrays.asList("breed 1", "breed 2"));
+        setupDataStoreWithUseCase();
+
+        subject.updateBreedList();
+
+        assertThat(subject.breedList).isEqualTo(breedListResponse.getBreedList());
+    }
+
+    private void setupDataStoreWithUseCase() {
+        FilterAnimalTypeUseCase useCase = mock(FilterAnimalTypeUseCase.class);
+        when(useCase.getAnimalType()).thenReturn(AnimalType.CAT);
+        when(dataStore.get(FilterAnimalTypeUseCase.class)).thenReturn(useCase);
     }
 }
